chore: consolidate resolvePort + drop bespoke port logic
Now that PR #90 reliably sets `window.__QONTINUI_PORT__` at webview boot,
the per-consumer port-resolution scaffolding can collapse to a single
source of truth.

Changes:
- Move `resolvePort()` from `LaunchMenu.tsx` to `src/lib/runner-api.ts`
  next to `getApiPort` / `getApiBase`. UI components shouldn't own API
  primitives.
- Drop `fileActivityApi.ts::apiBaseUrl()` — its bespoke `window` read +
  hardcoded `9876` fallback bypassed `getApiPort()` entirely, so it was
  always broken on temp/secondary runners (regardless of PR #90). Now
  uses the shared `resolvePort()`.
- Update import paths in the five consumers
  (`useFileLockTracking`, `useMidSessionProbe`, `useRegistryAwareness`,
  `useReviewState`, `useSessionManager`) + `LaunchMenu.test.tsx` to
  pull `resolvePort` from `@/lib/runner-api`.
- `buildProbeUrl`, `buildProbeBody`, etc. remain in `LaunchMenu.tsx` —
  they're probe-specific, not port-specific.

No behavioral change: `resolvePort()`'s body is identical to the old
one. All 53 existing tests (`LaunchMenu.test.tsx`, `FileActivityPanel.test.tsx`)
pass; tsc + eslint clean.

Session-Id: e962aa99-bad3-4091-8b7d-caa6491d8dfe
